change: Refactored theme decorator
